themes(tree): rework declaration and default-size
