+++ src/components/tree/themes/container.base.scss
@@ -1,7 +1,12 @@
 @use 'styles/common/component';
 @use 'styles/utilities/index' as *;
+@use './light/themes' as *;
+
+$theme: $bootstrap;
 
 :host {
+    --component-size: var(--ig-size, #{var-get($theme, 'default-size')});
+
     display: block;
     overflow: auto;
     font-family: var(--ig-font-family);
