+++ src/components/tree/themes/dark/container.indigo.scss
