+++ src/components/tree/themes/shared/item.material.scss
@@ -2,11 +2,9 @@
 @use 'styles/utilities' as *;
 @use '../light/themes' as *;
 
-$theme: $bootstrap;
+$theme: $material;
 
 :host {
-    --component-size: var(--ig-size, #{var-get($theme, 'default-size')});
-
     background: var-get($theme, 'background');
     color: var-get($theme, 'foreground');
 }
@@ -26,7 +24,7 @@ $theme: $bootstrap;
 }
 
 :host(:not([disabled])[active]) {
-    background:var-get($theme, 'background-active');
+    background: var-get($theme, 'background-active');
     color: var-get($theme, 'foreground-active');
 }
 
@@ -43,13 +41,9 @@ $theme: $bootstrap;
 :host(:not([disabled])[selected]),
 :host(:not([disabled])[selected][active]) {
     [part~='focused'] {
-        background: var-get($theme, 'background-active-selected');
+        background:var-get($theme, 'background-active-selected');
         color: var-get($theme, 'foreground-active-selected');
     }
-}
-
-[part~='select'] {
-    margin-inline-end: rem(8px);
 }
 
 [part~='focused'] {
